Fix up error handling
Initially identified with a confusing error message, this updates the error so that when the last person to edit the
page with a <pdf> tag in does not have the embed_pdf permission, an error is displayed where the pdf would be
displayed and the page is added to a tracking category.  This allows other users to come along and clean up the
problematic pages.

Note that the code did not stop users without permission from adding the tag previously, so that behavior is not
changed.

Fixes #20